SV_FatPVS(): use int32_t instead of long, as the code assumes it's 32bit

+++ src/client/cl_cin.c
@@ -417,7 +417,7 @@ SCR_ReadNextFrame(void)
 	FS_Read(&size, 4, cl.cinematic_file);
 	size = LittleLong(size);
 
-	if (((unsigned long)size > sizeof(compressed)) || (size < 1))
+	if (((size_t)size > sizeof(compressed)) || (size < 1))
 	{
 		Com_Error(ERR_DROP, "Bad compressed frame size");
 	}
